Extract firebase logic to mindbox-firebase module
