+++ settings.gradle
@@ -1,5 +1,6 @@
 include ':sdk'
 include ':facade'
 include ':mindbox-hms'
+include ':mindbox-firebase'
 include ':app'
 rootProject.name = "AndroidSdk"
